requirements.txt
beautifulsoup4==4.3.2
Django==1.6.5
django-ajax-selects==1.3.4
django-braces==1.4.0
django-ckeditor==4.4.7
django-ckeditor-updated==4.4.4
django-endless-pagination==2.0
django-grappelli==2.5.3
django-haystack==2.3.1
django-jsonfield==0.9.13
django-model-utils==2.0.3
django-shortcodes==0.3.0
django-social-share==0.2.1
django-tinymce==1.5.2
MySQL-python==1.2.5
Pillow==2.4.0
pysolr==3.3.0
pytz==2014.10
requests==2.3.0
six==1.7.3
South==0.8.4
unicode-slugify==0.1.3
Unidecode==0.4.17
xlwt==0.7.2
